+++ app/javascript/vue/tasks/digitize/store/mutations/TypeMaterial/setTypeMaterialBiologicalObjectId.js
@@ -1,3 +1,0 @@
-export default (state, value) => {
-  state.type_material.collection_object_id = value
-}